requirements.txt
PyQt5
numpy
opencv-python
websocket-client
rapidfuzz
pyaudio
pyttsx3
SpeechRecognition
sounddevice
python-dotenv
faster-whisper
onnxruntime
ultralytics
torch
piper-tts
